backend/notification-service/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
pymysql

pydantic
pydantic-settings
email-validator 

fastapi-mail
redis>=5.0.0

httpx 

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.2.0